Create base class for genomic validation methods
